added extra sanity checking
__init__ now checks for the presence of sourcefile for each chained
NOTES file.  all other needed files are implicitly checked because
they're needed to instantiate other objects.  also, moved the chdir
inside the looping try/catch block so that we can chdir(olddir) in a
finally statement.  this means we chdir for each NOTES file, but i
didn't have to worry about getting stuck in the wrong directly when
testing in the interpreter.