features: make register_feature track pre and post requirements
This commit is the 1st draft of an attempt to make register_feature track
not only the features that required before this one, but also the ones that
are required after it.  This is being accomplished by creating a registration
structure.

Note that this means the stage list is reduced to simply: build, install,
uninstall, and action.