notes: features can be enabled by adding NOTES section
This commit allows for implicit feature enabling by assuming that all
section headers correspond with a feature name (except for the default
sections).

In other words, if you add a [perms] section, it automatically adds perms
to the required feature list.

NOTE: This commit also changes the notes.options.flags map into the
      notes.options.features list because we're going to use feature
      sections in the NOTES file to pass arguments/variables into the
      various feature funcs.