features: added optional feature dependencies
This commit allows stage_structs to be created with features listed as deps
preceded with a question mark to indicate deps rules if a feature is
enabled, w/out recursively enabling it.  For example, if checksum is
enabled, strip_debug has to come first, but if it's not, strip_debug
doesn't care.