features/core: fleshed out install_func, v3
We're now installing manifest (which i'm going to go back and consistently
refer to as db) entry in /var/lib/srp/pkgname/SHA.  So far it's NOTES and
FILES.

Still a bunch of FIXMEs and half-implemented stuff...  upgrade logic, for
example is completely missing.  I should probably also go back and recreate
a fobj for the NOTES file at the very end of db entry creation so we can
update the NOTES file during installation.  And perhaps we should create
the db as yet another compressed tarball as apposed to a dir of files?