features: started plumbing for iter stages, v2
I've added a new test.py script here, trying to experiment with the
multiprocessing module.  I wanted to use the iter stages to transparently
feed work to parallel processes... which might still work out for most of
the work I was envisioning... but I don't think it'll work for iterating
over the tmp dir and adding files to the BLOB (tarfile).  Even if I could
get the proxy stuff to work right for accessing a tarfile via shared mem,
it would be pretty stupid.  A tarfile is really supposed to be a sequece of
bytes written to tape, so we would have to serialize all the add() calls...
which would defeat the purpose.

Maybe we just don't have the install_iter stage run in parallel... or maybe
we should make each Feature be able to specify whether each of its iter
stages can be run in parallel?