features: started plumbing for iter stages, v4
Added the FILES cached TarInfo list to brp.  But it doesn't work.
Apparently, you can't pickle a TarInfo object that was generated from a
TarFile object opened in write mode.  WTF?!  Testing in an interpreter with
a read-mode TarFile instance works fine...  Gonna have to rethink this a
little, because I was planning on letting the stage funcs store extra stuff
in the TarInfo objects' dicts... but they all exectute while we're
writing...

Actually, it could be the difference between the return value of
tar.getmember() vs tar.gettarinfo()...

Wait, TarInfo objects returned via gettarinfo() have a tarfile member that
needs to be removed before we can pickle it!  Adding a del(x.tarfile) call
to each item will make that list pickleable!