db: started actually coding, v3
Committing code from a few nights ago.  This is where I realized that NOTES
instances can't be pickled... which I think points out some of the
downsides of having the notes implementation be so dynamic.

I would like the db.installed_package constructor to generate a sha of the
notes and manifest objects, and I'd like to do that consistently across the
codebase by feeding the hashlib serialized data from pickle.dumps().  I had
actually been committing the notes instances into a file, and then reading
that into the hashlib, but that's just too complicated (and silly, I
think).