autotools: use pyexecdir instead of pythondir
At first I tried to install normal .py files in pkgpythondir and just the
compiled extension in pgkpyexecdir, but it looks like you really can't do
that.  The interpreter barfs if ANY part of a module was a C extension and
all .py and object files aren't in the same dir.