core: added functions file and FUNCTIONS env var, v3
Have to export CC and CXX.  In bash, specifying them on the command line
like that was exporting them but just for that line... In ash, it sets them
permanently but doesn't export.  The result is that the configure script,
which executes a couple sub-shells down, isn't seeing our CC/CXX variables.