notes: added a bunch of timing results to NotesBrp
I don't like how scattered this is, but it does add the following
statistics to the NOTES file instance in the finished brp:

time_total - total seconds it took to build/create the package.

time_build_script - seconds it took to finish running the build_script.

time_manifest_creation - seconds it took to create the manifest from the
     payload dir.

time_blob_creation - seconds it took to create the blob from the manifest
     (and payload dir).